Rework PBS and SGE, add tests
Create common base class for Grid Engine based cluster managers.

Add some tests, fix tests.
