+++ src/test/groovy/de/dkfz/roddy/execution/jobs/BEIntegrationTest.groovy
@@ -19,6 +19,8 @@ import groovy.transform.CompileStatic
 import org.junit.AfterClass
 import org.junit.BeforeClass
 import org.junit.Test
+import org.xml.sax.SAXParseException
+
 import java.time.Duration
 
 /**
